requirements.txt
asttokens==2.4.1
attrs==23.2.0
certifi==2024.7.4
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
executing==2.0.1
fastjsonschema==2.20.0
fonttools==4.49.0
geopandas==1.0.1
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kaleido==0.2.1
kiwisolver==1.4.5
matplotlib==3.8.3
matplotlib-inline==0.1.7
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==1.26.4
packaging==23.2
pandas==2.2.2
parso==0.8.4
pillow==10.2.0
platformdirs==4.2.2
plotly==5.23.0
prompt_toolkit==3.0.47
psutil==6.0.0
pure_eval==0.2.3
pyaml==23.12.0
Pygments==2.18.0
pyogrio==0.9.0
pyparsing==3.1.1
pyproj==3.6.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
rpds-py==0.19.1
scipy==1.12.0
shapely==2.0.3
six==1.16.0
stack-data==0.6.3
tenacity==9.0.0
tornado==6.4.1
traitlets==5.14.3
tzdata==2024.1
wcwidth==0.2.13
geopandas==1.0.1
multipledispatch==1.0.0
asttokens==2.4.1
attrs==23.2.0
certifi==2024.7.4
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
executing==2.0.1
fastjsonschema==2.20.0
fonttools==4.49.0
geopandas==1.0.1
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kaleido==0.2.1
kiwisolver==1.4.5
matplotlib==3.8.3
matplotlib-inline==0.1.7
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==1.26.4
packaging==23.2
pandas==2.2.2
parso==0.8.4
pillow==10.2.0
platformdirs==4.2.2
plotly==5.23.0
prompt_toolkit==3.0.47
psutil==6.0.0
pure_eval==0.2.3
pyaml==23.12.0
Pygments==2.18.0
pyogrio==0.9.0
pyparsing==3.1.1
pyproj==3.6.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
rpds-py==0.19.1
scipy==1.12.0
shapely==2.0.3
six==1.16.0
stack-data==0.6.3
tenacity==9.0.0
tornado==6.4.1
traitlets==5.14.3
tzdata==2024.1
wcwidth==0.2.13
geopandas==1.0.1
multipledispatch==1.0.0
area==1.1.1
